Use CustomCastMediaManager to allow for casting mid-playback
When connecting cast mid-playback, the SDK loads MediaInfo to the
receiver automatically, which was bypassing our custom data logic.

By creating our own CustomCastMediaManager class that adds the
licenseHeaders custom data with load and add calls, the MediaInfo
is always correct and plays successfully, regardless of where it
is called from.